moved config files into src, renamed backrefs for relationships
