--- .env
@@ -1,2 +1,0 @@
-DATABASE_URL='postgresql+psycopg2://admin:adm1n@localhost:5000/fitnessfinder_db'
-SECRET_KEY="FF2023"
