+++ src/.env
@@ -1,0 +1,2 @@
+DATABASE_URL='postgresql+psycopg2://admin:adm1n@localhost:5432/fitnessfinder_db'
+SECRET_KEY="FF2023"
