+++ src/.flaskenv
@@ -1,2 +1,4 @@
 FLASK_APP=app:create_app
 FLASK_ENV=development
+FLASK_RUN_PORT=5000
+
