+++ src/models/facilities.py
@@ -28,21 +28,12 @@ class Facility(db.Model):
     address_id = db.Column(db.Integer, db.ForeignKey("addresses.id"), nullable=False)
     
     # Add the relationships directions to other models
-    promotions = db.relationship('Promotion', backref='facilities', lazy=True, cascade="all, delete-orphan")
+    promotions = db.relationship('Promotion', backref='facility_promotions', lazy=True, cascade="all, delete-orphan")
     address = db.relationship('Address', backref='facility')
-    owner = db.relationship("Owner", backref=db.backref("facilities", cascade="all, delete-orphan"))
-    facility_type = db.relationship('FacilityType', backref=db.backref('facilities', lazy=True))
+    facility_type = db.relationship('FacilityType', backref='facility_types', lazy=True)
+    owner = db.relationship("Owner", backref="owner_facilities", cascade="all, delete-orphan", single_parent=True)
+
 
     # Set up join table with amenities
     facility_amenities = db.relationship('Amenity', secondary=facility_amenities, lazy='subquery', backref=db.backref('facilities', lazy='dynamic'))
 
-
-    def __init__(self, phone_num, independent, business_name, opening_time, closing_time, facility_type, owner_id, address_id):
-        self.phone_num = phone_num
-        self.independent = independent
-        self.business_name = business_name
-        self.opening_time = time(hour=opening_time.hour, minute=opening_time.minute)
-        self.closing_time = time(hour=closing_time.hour, minute=closing_time.minute)
-        self.facility_type = FacilityType.query.filter_by(facility_type=facility_type).first()
-        self.owner_id = owner_id
-        self.address_id = address_id
